requirements.txt
# requirements.txt
aiohttp>=3.8.0
python-telegram-bot>=20.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
aiofiles>=23.1.0
cachetools>=5.3.0